Add python checker 'python'
The interpreter itself is able to detect big problems, e.g. SyntaxError
or problems in top-level code. Therefore it makes a good first-pass
checker.

The python command-line must be passed some code (~10 statements) in
argument, which is done a bit inelegantly here, but doc on multi-line
strings in vim script is quite rare...