Avoid changing the args list of the original maker.
If a dictionary contains a list then the items in the list are stored as references. This means that the call to expand on line 66 expands the filename that is stored in the original list in the original maker, this in turn means that once MakeJob has been called for the first time the string literal '%:p'  will never appear in the args list. This leads to line 63 always evaluating true and the file name being appended to the args list each time MakeJob is called.