Experiment with several flow errorformats
The current flow maker errorformat is being flagged as invalid and thus being
removed from the location list (#27).  This results in Vim just saying that the
maker exited with an error code, but not actually listing any errors.  In an
attempt to display errors in the location list, I'm experimenting with different
errorformats.

Flow can output different error strings depending on the arguments passed, one
of which is '--from=vim' which is a little easier to parse for Vim than the
standard output.  However, the resulting message is arguably too cryptic to
understand without more context, hence attempts to try differnet formats.