RFC: Annotate lines with the maker name
This is useful (only) with multiple makers, which is not handled
currently.
For this, it should get done in AddExprCallback probably (once `index`
is > 1), but this then needs an extra list to keep track of maker names
per line (it does not seem to be possible to store extra data with the
location/quickfix list itself (`entry.maker_name`)).