Process job output by window instead of buffer
This is a rather huge refactor.

Fixes https://github.com/benekastah/neomake/issues/350 and
https://github.com/benekastah/neomake/issues/348.

It removes 'winnr' from 'maker'.

The behaviour changes so that the output is only processed by windows,
but not buffers.  This means that `Neomake | sp` will not open the
location list, until you go to the previous window.