Only handle CursorMoved when cursor was moved between lines
Moving cursor left and right staying on the same line causes lags,
because visible signs and error messages are redrawn on every move.

Looks like there is no need to do it, because the first error message
for the line is shown anyway, and visible area is not changed.

This PR is my naive attempt to avoid this lag by handling `CursorMoved`
only when line under cursor was changed.

Closes https://github.com/neomake/neomake/pull/394.