Adds support for typescript projects.
Right now we have a typescript file maker that builds a single file
with tsc.  This commit adds support for "typescript projects" which
run the typescript compiler on every typescript file contained in
a typescript project. (a typescript project is a folder that has
a tsconfig.json file somewhere in the directory hierarchy).

I also made a bugfix in neomake.vim which wasn't handling the
non-existance of bufnumber that could happen when running a maker
without having any files opened in neovim.