Do not process current window when in quickfix buffer
This cannot be done in a reliable way easily (we would have to compare
the jobinfo winnr/bufnr, and should use winids then), and it is not
really worth it.  After all I consider it even disturbing if the
location/quickfix list would get changed while you are browsing/using it.