Improve managing active jobs, and restart them directly
This manages active jobs in s:make_info, and keeps track of finished
ones.
It now restarts jobs for the same maker directly, instead of waiting for
the old one to do it from the exit handler.

This fixes several issues (new tests), and makes the processing more
reliable/sane hopefully.