Add Elm maker using elm-make
This adds a maker for [Elm](http://elm-lang.org/). The compiler
errors provided by elm are quite detailed, but this only adds
handling for the base error/warning types and short messages.

This maker relies on the JSON output of elm-make and is *heavily*
based on #942 (it's basically a carbon copy).