Add dictionary based config
This will lookup config settings in b:neomake, t:neomake, g:neomake,
and has support for `context` (for filetype and bufnr) (where
`context.maker` is also looked at in the end).

This is considered to be experimental (and is not documented therefore).
`neomake#utils#GetSetting` will look at the new-style config first.

Some functions (like `neomake#config#get_with_source`) are not used yet,
but are meant to be with the automake feature branch.