Fix bufnr handling with get_list_entries/process_output
`bufnr` does not default to the job's bufnr anymore, but is expected to
be given as `bufnr` or `filename`.  The `bufnr` is then picked up from
the parsed location/quickfix list entries (where Vim will have created
an unlisted buffer for non-opened filename entries).

Change to job cwd in s:ProcessEntries, and when filling the
location/quickfix list.

Adds cwd related tests:

- process_output: filename with cwd
- legacy errorformat maker: filename with cwd
- legacy errorformat maker: filename with cwd (error: removed)
- legacy errorformat maker: filename with cwd (error: maker rmdir)

Ref: https://github.com/neomake/neomake/pull/1194#pullrequestreview-35487107
Ref: https://github.com/neomake/neomake/pull/1218