Fix closing of empty location/quickfix lists
Fixes https://github.com/neomake/neomake/issues/122#issuecomment-303892087.
Regressed in 07ec3327.

Without patch-7.4.379 it could create a new list to work around
`:lwindow` opening it for invalid entries, but that seemed to be too
much of a workaround.
In the end we would probably filter the list ourselves and use
:lopen/:copen instead.

Closes https://github.com/neomake/neomake/pull/1371.