Add neomake#utils#fix_self_ref
Add a helper function to fix self-references in objects, and use it when
logging remaining jobs in tests.
This is required in case a maker stores the jobinfo object on `self`
etc.

Fixes in Vim patch 7.4.1644, but this function is still useful, since it
uses named refs instead of just "...".