Getting CLASSPATH from Apache Ant build file. (#1544)
To make this work, you need a task called `classpath` on the Ant
build.xml file, that returns the classpath needed to compile the current
file on Vim. It can be a fixed string or a command that returns a
combination of paths, like:

```
<?xml version="1.0"?>
<project name="neomake" default="build" basedir=".">
   <path id="compile.path">
      <pathelement location="a.jar"/>
      <pathelement location="b.jar"/>
      <pathelement location="c.jar"/>
   </path>

   <target name="classpath">
      <echo message="${toString:compile.path}"/>
   </target>
</project>
```

which will print:

```
$ ant classpath -f build.xml -S -q
/tmp/a.jar:/tmp/b.jar:/tmp/c.jar
```
that can be used on Neomake.