C and C++ maker improvements (#1488)
* Added path to clang-tidy maker

* clang-tidy and clang-check use clang/clang++, which must be config'd

We are now passing the same parameters we would pass to clang to these
tools, separated from the actual tool parameters with a "--" argument.

* Include the current (root/project) directory in cpp search path

For single-file checkers for CPP (both clang and gcc), include the
current directory in the search path. The project makers will always
need to be configured, but this should help with the default
configuration for single-file makers.

* Moved single-file include search path setting to the C makers

Instaed of adding -I./ to gcc/clang for cpp only, now adding them
for both c and cpp filetypes by specifying them in the default c maker
configuration.

* Removed clang-tidy/clang-check inheritance from cpp maker

The -fsyntax-only parameter was causing clang-tidy and clang-check
to disregard the remaining clang arguments (after --) for some reason.