loclist/qflist: use number of entries for window height
This will take into account the number of list entries (e.g. 3) when
creating a location list or quickfix window, and uses the minimum of
`g:neomake_list_height` and that.

Therefore with only 3 entries it will use 3 instead of 10 now by
default.

It is still a good idea to use something like
https://github.com/blueyed/vim-qf_resize in general, but this improves
the default behavior.