Enhanced quickfix/loclist via annotations
You have to call `neomake#quickfix#enable()` for now to enable it.

There is missing coverage, but it is opt-in and experimental, and I
think it is good to build upon it from the current state - given that it
is almost a year old and was not trivial to rebase anymore.

Thanks a lot to Tommy Allen for proposing this initially!

Closes https://github.com/neomake/neomake/pull/924.