python: flake8: fix length processing in f-strings
"F821 undefined name 'bar'" with `f'foo bar'` refers to the beginning of
the f-string, and therefore the entry's `length` property should not be
set.

This changes the flake8 postprocessing to use the generic length
postprocessor, which results in no `length` attribute for this entry
(since it checks if the text is there).

A proper fix would be to search for the next `{bar}`, and adjust the
lnum/col also.