fix failing test and better organize its files
Fix failing test, since python detects error as "parenthesis never
closed" instead of "unexpected end of line". As such, the error is now
detected in the '(' character, not in the '\n' caracter so decrement the
column number by one.

Also move the fixture to a specific directory, so it is more organized